backend/requirements.txt
Django
djangorestframework
django-cors-headers
python-dotenv
psycopg[binary]
pgvector

google-generativeai
mistralai
langchain-text-splitters
rapidocr-onnxruntime
PyMuPDF
PyPDF2
Pillow
unidecode
